requirements.txt
amqp==2.6.1
asgiref==3.3.1
autopep8==1.5.5
billiard==3.6.3.0
boto3==1.17.19
botocore==1.20.19
celery==4.4.7
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.7
djangorestframework==3.12.2
dnspython==1.16.0
docopt==0.6.2
docutils==0.16
fett==0.3.2
gevent==21.1.2
greenlet==1.0.0
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
kombu==4.6.11
networkx==2.5
pycodestyle==2.6.0
pymongo==3.11.3
python-dateutil==2.8.1
python-decouple==3.4
python-jsonrpc-server==0.3.4
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
snooty-lextudio==1.8.10.dev0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
ujson==1.35
urllib3==1.26.3
vine==1.3.0
watchdog==1.0.2
zope.event==4.5.0
zope.interface==5.2.0
